making a commands folder and doing some not great OOP
