+++ src/Aspire.Cli/README.md
@@ -1,0 +1,12 @@
+## Aspire.Cli Proposed Surface
+
+Root command: **aspire**
+
+Global options for debugging:
+`-d, --debug`
+`-w, --wait-for-debugger`
+
+Subcommands:
+```cli
+aspire run [--project <PATH_TO_CSPROJ>] 
+```
